+++ src/lib/forms-schema/property-schemas.ts
@@ -63,7 +63,7 @@ export const requiredSchemas = {
 const readOnly = Joi.bool().default(false)
 
 const approverEditability = Joi.object({
-  type: Joi.valid('ALL_STEPS'),
+  type: Joi.valid('ALL_STEPS').required(),
 })
 
 export const readOnlySchemas = {