+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `approverEditability` on form elements so they can be edited by approvers
+
 ## [13.2.3] - 2026-08-26
 
 ### Added